Restrict Pages

+++ base/views.py
@@ -79,6 +79,11 @@ def CreateRoom(request):
 def UpdateRoom(request,pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    messages.error(request,'You are not allowed to do this!')
+    return redirect('home')
+
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
     if form.is_valid():
@@ -90,6 +95,9 @@ def UpdateRoom(request,pk):
 @login_required(login_url='login')
 def DeleteRoom(request,pk):
   room = Room.objects.get(id=pk)
+  if request.user != room.host:
+    messages.error(request,'You are not allowed to do this!')
+    return redirect('home')
   if request.method == 'POST':
     room.delete()
     return redirect('home')
